# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "PikPakGo API",
        "description": "PikPakGo Travel Marketplace API - Hotels, Vacation Rentals & More",
        "contact": {
            "name": "PikPakGo Support",
            "email": "reservations@pikpakgo.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local Development Server"
        },
        {
            "url": "https://pickpackgo.in-sourceit.com/api",
            "description": "Production Server"
        },
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/wishlist": {
            "get": {
                "tags": [
                    "Wishlist"
                ],
                "summary": "Get current user's wishlist",
                "operationId": "3623a3b40a4fb4addae4953160ec84dc",
                "responses": {
                    "200": {
                        "description": "Wishlist items"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Wishlist"
                ],
                "summary": "Add a property to wishlist",
                "operationId": "d632b05deecee49d31504522e14560d4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "property_code",
                                    "property_name"
                                ],
                                "properties": {
                                    "property_code": {
                                        "type": "string",
                                        "example": "orp5b27f9x"
                                    },
                                    "property_name": {
                                        "type": "string",
                                        "example": "Beach House"
                                    },
                                    "provider": {
                                        "type": "string",
                                        "example": "ownerrez"
                                    },
                                    "property_city": {
                                        "type": "string",
                                        "example": "Miami"
                                    },
                                    "property_country": {
                                        "type": "string",
                                        "example": "US"
                                    },
                                    "featured_image": {
                                        "type": "string",
                                        "example": "https://..."
                                    },
                                    "price_from": {
                                        "type": "number",
                                        "example": 150
                                    },
                                    "currency": {
                                        "type": "string",
                                        "example": "USD"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Added to wishlist"
                    },
                    "200": {
                        "description": "Already in wishlist"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/wishlist/{propertyCode}": {
            "delete": {
                "tags": [
                    "Wishlist"
                ],
                "summary": "Remove a property from wishlist",
                "operationId": "c369fa145c6c57437f942431957c5823",
                "parameters": [
                    {
                        "name": "propertyCode",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Removed from wishlist"
                    },
                    "404": {
                        "description": "Not in wishlist"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/wishlist/check/{propertyCode}": {
            "get": {
                "tags": [
                    "Wishlist"
                ],
                "summary": "Check if a property is in the user's wishlist",
                "operationId": "b62d99f2f6693db059573daf3e40ecc0",
                "parameters": [
                    {
                        "name": "propertyCode",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Saved status"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}